some more changes

+++ src/background/queue_item.rs
@@ -9,7 +9,7 @@ pub struct QueueItem {
     pub queue_item_contents: String,
     pub work_started: Option<i64>,
     pub being_worked: bool,
-    pub error_count: i64,
+    pub error_count: i32,
     pub error_message: Option<String>,
 }
